requirements.txt
# mysqlclient==1.3.10
Django==1.11.3
djangorestframework==3.6.3
django-cors-headers==2.1.0
pre-commit==0.15.3
elasticsearch-dsl==5.3.0
python-memcached==1.58
pandas==0.20.3
